Many improvements to the README.
I removed the LICENSE.txt file because our own README points to SciRuby's
license, not to it -- it was basically occupying space in the repository.

Someone sent me an e-mail with some complaints and suggestions about the
README and the documentation. For example, I added links to BLAS, CBLAS, LAPACK
and ATLAS, as many people don't know what they are (I didn't when I started
contributing).

"Requirements" was moved to "Installation", I added "Documentation" and fixed
some problems in "Examples". I removed the list of future features and copied
it over to the Roadmap in the wiki, where it belongs. Of course, I added a link
in the README.

Apart from that, I added several links to the issue tracker, the mailing list, the installation page of the wiki and to the documentation at sciruby.com.

It seems pretty neat now.